feat: add runtime health reset and deterministic tests
