+++ apps/api/src/tests/hardening.spec.ts
@@ -1,27 +1,45 @@
-import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { beforeEach, afterEach, describe, expect, it, vi } from 'vitest';
 
 let buildServer: typeof import('../server.js').buildServer;
 let setJobBeat: typeof import('@prism-apex-tool/runtime').setJobBeat;
+let __resetHealth: typeof import('@prism-apex-tool/runtime').__resetHealth;
 
 beforeEach(async () => {
   vi.resetModules(); // avoid double registration/state
+  vi.useFakeTimers();
   process.env.RATE_LIMIT_MAX = '2';
   process.env.RATE_LIMIT_WINDOW_MS = '60000';
   process.env.RATE_LIMIT_MAX_BUCKETS = '10';
   delete process.env.BEARER_TOKEN; // ensure auth is OFF for these tests
   ({ buildServer } = await import('../server.js'));
-  ({ setJobBeat } = await import('@prism-apex-tool/runtime'));
+  ({ setJobBeat, __resetHealth } = await import('@prism-apex-tool/runtime'));
+  __resetHealth();
+});
+
+afterEach(() => {
+  vi.useRealTimers();
 });
 
 describe('Hardening', () => {
   it('returns readiness', async () => {
     const app = buildServer();
+
+    // Set all jobs to healthy
+    setJobBeat('alpha');
+    setJobBeat('beta');
     setJobBeat('marketFeed');
+
+    // No real sleeping; the health uses Date.now, so it's instant
+
     const res = await app.inject({ method: 'GET', url: '/ready' });
     expect(res.statusCode).toBe(200);
     const body = res.json();
-    expect(body.jobs.marketFeed.healthy).toBe(true);
-    expect(body.overall).toBe('healthy');
+
+    // Ensure job is healthy and overall is healthy
+    expect(body.jobs.alpha?.healthy).toBe(true);
+    expect(body.jobs.beta?.healthy).toBe(true);
+    expect(body.jobs.marketFeed?.healthy).toBe(true);
+    expect(body.overall).toBe('healthy'); // Expect overall to be healthy
     await app.close();
   });
